Sound fixes for block placing/breaking (#10423)
When placing, hitting, or breaking a block, the same pitch and volume calculations are used as in Vanilla functions.
In practice: block sounds are lowered in pitch to match Vanilla sounds.
When a builder places a solid substitution block, use the sound effect of the block that is substituted instead of the sound of the solid substitution block itself.
In practice: builders placing dirt blocks to fill an area now makes dirt placing sounds instead of wood placing sounds.
When hitting a block but not yet breaking it, use the block's hit sound instead of the break sound.
In practice: a builder digging up e.g. grass blocks now makes a lot less noise.