Lumberjack handling of productive bees (#10944)
Generalizes the previously existing Mangrove pattern to all base trees (allowing a family of logs to be recognized as part of the same tree), and adds the productive bee nests as a a recognized part of the tree family.
As a result bee nests of the same base type as the logs in their tree will be harvested, but a hand-placed nest in a different tree (say, an oak nest in a spruce tree) will be left behind intentionally.