Safe removal of add-on jobs and buildings. (#11032)
Adjust the error handling of restoring a saved game such that if an add-on mod is removed (say... Trade Post) the colony doesn't become corrupted by the unrecognized buildings and jobs that it included. This will log the errors, but just leave the citizens unemployed instead of corrupting the load process.